add setup file
