+++ src/setup.ts
@@ -1,0 +1,16 @@
+import { serverSetup } from '@graphql-schema/connection';
+import * as express from 'express';
+
+export const start = async (): Promise<express.Express> => {
+  const app = express();
+  await serverSetup(app);
+  return app;
+};
+
+export const listen = async () => {
+  const app = await start();
+  const PORT = process.env.PORT ?? 3000;
+  app.listen(PORT, () => {
+    console.info(`Server is running on port ${PORT}`);
+  });
+};
